requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
fastapi==0.119.0
h11==0.16.0
google-generativeai==0.7.2
httptools==0.7.1
idna==3.11
narwhals==2.8.0
numpy==2.3.3
packaging==25.0
plotly==6.3.1
pydantic==2.12.1
pydantic_core==2.41.3
python-dotenv==1.1.1
PyYAML==6.0.3
scipy==1.11.4
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
